fio: Added new package.
fio is a tool to simulate a given workload to test for performance.
Fio spawns a number of threads or processes doing a particular type of
I/O action as specified by the user. The typical use of fio is to write
a job file matching the I/O load one wants to simulate.

Change-Id: I8966c815594af7aeb4c11405395621e011af62ff
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/16769
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Harinadh Dommaraju <hdommaraju@vmware.com>