fio: Added new package.
fio is a tool to simulate a given workload to test for performance.
Fio spawns a number of threads or processes doing a particular type of
I/O action as specified by the user. The typical use of fio is to write
a job file matching the I/O load one wants to simulate.

Change-Id: Ib705eae2088c49c222e7769889e0ef0e9905ac1f
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/17110
Reviewed-by: Michelle Wang <michellew@vmware.com>
Tested-by: Piyush Gupta <gpiyush@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>