stalld: Exit early if enabling HRTICK fails when using SCHED_DEADLINE
stalld allows the boost runtime that can be specified to be at a
maximum of 1ms. At this granularity, the HZ tick will not be enough to
make scheduling decisions at the needed time. Therefore HRTICK needs
to be enabled to ensure that SCHED_DEADLINE tasks only run for a duration
of "dl->runtime" and not end up overrunning their allowed bandwidth.

If HRTICK cannot be enabled for some reason, we should not be allowing
stalld to run in a mode where it is doing priority boosting using
SCHED_DEADLINE. Add the necessary code for this.

We also need to enable HRTICK only when stalld is using SCHED_DEADLINE.
If stalld is using SCHED_FIFO for priority boosting, it mimics SCHED_DEADLINE
using calls to nanosleep. The nanosleep system call sets up its own hrtimers
independent of HRTICK.

Change-Id: Ib8e98b18e1e9a7f6c809fa11a59a73e720d11296
Signed-off-by: Sharan Turlapati <sturlapati@vmware.com>
Signed-off-by: Keerthana K <keerthanak@vmware.com>
Tested-by: Sharan Turlapati <sturlapati@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/17079
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>