unbound: Fix for CVE-2022-30698 and CVE-2022-30699
Change-Id: I08a703763c18c67e2d0a0f39046a852c6dfcdea1

This CVE-fix consists of 2 fixes.

Patch 1: 0001-Fix-that-nxdomain-synthesis-does-not-happen-above-th.patch

This patch is a dependency for fixing CVE-2022-30698 and CVE-2022-30699. Two pointers, "delname" and "dpname" were introduced in this patch, which were being used by the next patch (0001-unbound-Fix-for-CVE-2022-30698-and-CVE-2022-30699.patch)

Patch 2: 0001-unbound-Fix-for-CVE-2022-30698-and-CVE-2022-30699.patch

This patch fixes CVE-2022-30698 and CVE-2022-30699. Some of the files did not get patched correctly due to some hunk failures. These files had to be tweaked individually with very minor changes without altering the logic of the code. These files are 
1. doc/Changelog 
2. iterator/iter_utils.h
3. pythonmod/pythonmod_utils.c

Change-Id: I08a703763c18c67e2d0a0f39046a852c6dfcdea1
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/17512
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>