kpatch: adding rpm packaging capability and modinfo visibility
This commit introduces a feature to package livepatch modules as RPMs.
It supports setting RPM version and release, and optionally a separate
description from the livepatch module.

Also included is a patch for the kpatch utility that changes the installation
directory to one that is visible to modinfo. Previously, livepatch modules would
show up in lsmod, but would not show up with modinfo. This fixes that, and allows
modinfo to print information about installed livepatch modules without compromising
any kpatch installation functionality.

Change-Id: Ib081fe13ff160ed037214e29c28bb10323cf8900
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/17795
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>