linux-esx: ptp_vmw driver updates
Extend ptp_vmw driver's functionality to allow a VM to make time
adjustments to the PHC device, enabling mechanisms that allow
time sync agents running in a VM to provide time to other VMs
and the underlying host. Also add a module parameter probe_hv_port
that allows ptp_vmw driver to be loaded even when ACPI is disabled.

Change-Id: Ib38d6127e59701bf6263d5207a210c8c04209ed0
Originally-by: Deep Shah <sdeep@vmware.com>
Signed-off-by: Deep Shah <sdeep@vmware.com>
Reviewed-by: Vivek Thampi <vithampi@vmware.com>
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/17852
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Reviewed-by: Ajay Kaher <akaher@vmware.com>