linux-rt: Add guest timer padding
This is an optimization to hide the virtualization cost of timer
interrupts.

The idea is that if we have a predictable overhead, we can prepone the
timer by the said overhead. As a result we would receive the interrupt
in the guest at the same time as baremetal, giving the impression that
there is no virtualization overhead.

Next we have to determine what value to prepone the timer by.  One of
the fundamental guarantees of timer interrupts is that a timer's
callback will never be invoked before its configured time.  If the
timer is preponed, it needs to be ensured that this guarantee is
satisfied.  In order to do that, the interrupt handler needs to spin
in the guest, in case the timer is received earlier than intended.
This is not a desirable scenario, especially in a real-time system, as
this will be hogging the CPU from the workload, with interrupts
disabled.

When we traced the observed overhead from the guest for timer
interrupts, we found that in about 99.9% of the instances the overhead
was within a range of about 400ns, from approx. 1200ns to 1600ns.

So we decided to use the minimum observed overhead since boot time to
prepone the timer.  The miniumum observed overhead is continuously
monitored during runtime.  With this option, we are minimizing the
need to spin while also not giving up too much of the possible latency
gains.

Change-Id: I0313c4dcc0eeb68625a43913c86852b6ffc83dac
Originally-by: Him Kalyan Bordoloi <bordoloih@vmware.com>
Signed-off-by: Him Kalyan Bordoloi <bordoloih@vmware.com>
Signed-off-by: Keerthana K <keerthanak@vmware.com>
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/17978
Reviewed-by: Ajay Kaher <akaher@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>