abupdate: initial addition
Initial addition of abupdate package/script. Contains script, systemd service
for auto finalizing, and a configuration file.

This is designed to facilitate A/B system upgrades, where the update/modification
occurs on the inactive (B) partition set and is only activated by switching partition sets.
More details should be in the README

Change-Id: I3b0229d06a2c228649ec4b237672c209d59cb0ae
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/18514
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>