linux-rt: Add multi-version support for i40e, iavf and ice drivers as sub-packages
Add support to provide multiple versions of the i40e, iavf and ice
drivers as sub-packages of linux-rt.

Design Considerations:
=====================

1. Package Upgrades

Normally, when carving out existing functionality of a package into
its own sub-package (eg: moving ice.ko.xz from linux-rt to its
sub-package linux-rt-drivers-intel-ice-1.8.3), we add a 'Requires'
dependency to the main package, such that updating an older
installation of the main package automatically brings in the
sub-package as well (thus making sure that the package upgrade doesn't
break or lose existing functionality).

However, in this case, we are not adding this 'Requires' dependency
from the linux-rt main package to each of the driver sub-packages
being created, due to certain limitations of the project requesting
these drivers.

Therefore, the user is expected to also specify the driver
sub-packages they need (if required) when updating from older versions
of linux-rt.


2. Installing drivers when the usecase is version-agnostic (latest)

Since we are packaging specific versions of the drivers as
sub-packages, the sub-package names include those version numbers too
(eg: linux-rt-drivers-intel-ice-1.8.3). This helps support usecases
that need to install those specific driver versions.

However, to simplify usage scenarios where only the latest driver is
needed (i.e., version-agnostic), we add a 'Provides' tag to the latest
version of each driver with a capability name that does not include
the version info.

Eg: linux-rt-drivers-intel-ice-1.8.3 (which is currently the latest
ice driver), can also be installed by specifying
linux-rt-drivers-intel-ice.


3. Common Dependencies

iavf v4.4.2 and ice v1.8.3 drivers both need the auxiliary.ko.xz
kernel module. So we include the auxiliary module in the main linux-rt
package to simplify dependencies between these drivers.


4. Preventing simultaneous installations of multiple driver versions

To prevent users from installing multiple versions of the same driver
simultaneously, we add 'Conflicts' tags to each driver version
sub-package to make it mutually-exclusive with its other versions.

Change-Id: I842e1bd44ded0ca2517e906eff78feb649d056dc
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/18586
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Ankit Jain <ankitja@vmware.com>
Reviewed-by: Keerthana K <keerthanak@vmware.com>