openssl: tweak openssl.cnf
Keep default provider enabled at all times. This is needed to
momentarily allow non fips algos if needed.

An upstream example of such usecase:
https://github.com/MariaDB/server/pull/2010/files

A simple program with md5 usage like in the above example should work in
our environment. And `openssl md5 <file>` should fail which implies that
fips is the default provider.

In this patch it's also taken care not to modify the openssl.cnf which
will endup in unexpected situation during openssl upgrades.

base_sect is redundant when default provider is enabled, hence removed
it from provider_fips.cnf

Change-Id: I6a7f21a5f2193ae2bccf656c19ff7edffb51d54f
Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/19120
Tested-by: Michelle Wang <michellew@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>