stalld: Backport fix to detect and boost runnable dying tasks
In some scenarios, a process which is dying (X state) can end up not
making any progress due to CPU starvation. And this in turn can cause
other threads in its thread group to deadlock too.

So, backport a fix from upstream stalld to detect and boost tasks that
are starved in the dying state.

Change-Id: Ia1501e50f2186f9576befa2bfb232ab26b2343fc
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/19418
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Keerthana K <keerthanak@vmware.com>