linux-rt: fix regression in rt patchset that leads to deadlock/system hang
Discovered a regression in the 4.19.255-rt114 patchset, in
0347-workqueue-Use-rcuwait-for-wq_manager_wait.patch, which was consumed
when we upgraded linux-rt to version 4.19.268-1 (RT patchset rt117).
This regression has the ability to cause complete system hangs due to
a deadlock in critical kernel code.

It seems that the other Linux RT branches backported this patch correctly,
but in the 4.19 branch there was a mistake.

This regression is caused by the failure to remove the acquisition of
pool->lock in put_unbound_pool() in kernel/workqueue.c. pool->lock
is also acquired in wq_manager_inactive(), which is called in the
immediate next line. We can remove the first lock acquisition, which
should have been removed in the backport anyways, to avoid recursive
locking and deadlock issues.

Change-Id: I293972b504901001e0a9907fd4b3d95100b950a0
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/19435
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>