linux, tools: Enable turbostat to work in the guest on VMware hypervisor
Turbostat hasn't really worked full-fledged within Linux guests on
ESXi, since some of the MSRs that it needs in order to gather data
have traditionally not been exposed to VMs by ESXi.

This is going to change in newer versions of ESXi, where these MSRs
will be accessible from guests for read operations. In preparation to
get turbostat working in the guest with these underlying changes, skip
some CPUID checks while accessing certain MSRs when running on VMware
hypervisor.

(Note that some of the data reported by turbostat will only be
valid/reliable if the guest has its vCPUs exclusively pinned to pCPUs,
because the data actually pertains to the underlying physical CPU,
such as the CPU frequency, thermal and power statistics.)

Change-Id: Icc7c06f577f47f112f871e79fb64884a536349f0
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/19758
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>