stalld: Backport fix to detect and boost runnable dying tasks
In some scenarios, a process which is dying (X state) can end up not
making any progress due to CPU starvation. And this in turn can cause
other threads in its thread group to deadlock too.

So, backport a fix from upstream stalld to detect and boost tasks that
are starved in the dying state.

Change-Id: I028e314ccf1ee215475ba6572878956f8dbdb587
Signed-off-by: Srivatsa S. Bhat (VMware) <srivatsa@csail.mit.edu>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/19759
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Him Kalyan Bordoloi <bordoloih@vmware.com>