openssl: package certified fips.so in openssl-fips-provider
When openssl is built with 'enable-fips' flag, it builds the latest
version of fips.so (fips-provider) which might not be certified.
Hence remove 'enable-fips' flag and use the certified fips-provider
version and package it in openssl-fips-provider.

Add a provision to build either latest or certfied fips.

Add some more strict flags to gcc while compiling jitterentropy and fix
reported warnings.

Fix various security issues.

Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>
Change-Id: I455ebc0f75116c9f8033a312eb7fc7604f69e98c
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/19787
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>