tuned: detect dbus config after installation
The tuned daemon fails to start after installation in some cases. This
seems to be due to the dbus daemon not detecting the new config file
added by tuned. Tuned puts the dbus config file in /etc/dbus-1/system.d,
which seems to have been deprecated as a location for third party packages.
We should move this config file to /usr/share to be in line with latest dbus

Change-Id: I425f6d1726645e39c04bc56ca6b35d4502e9f264
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/19790
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>