linux, linux-rt, linux-esx: clean up intel ethernet driver installation
The previous commit updating the intel ethernet driver versions was messy,
as some workarounds had to be left in to avoid the problem of the auxiliary
driver module getting installed on "modules_install" by every different driver
version that has it (not all driver versions compile it).

Fix this messiness by adding patches to the driver Makefiles to not install
auxiliary.ko/intel_auxiliary.ko (aux module) when using
"modules_install_no_aux".

Some driver versions do not compile the aux module, so in those cases just
add an alias to map "modules_install_no_aux" to "modules_install".
These patches are labeled as:
"<driver-version>-add-alias-for-modules_install_no_aux.patch"

Other driver versions include the auxiliary module, but have no build
target for "modules_install_no_aux". For these, backport this build target,
including the patch to not install the aux module.
These patches are labeled as:
"<driver-version>-introduce-modules_install_no_aux.patch"

The rest of the driver versions both include the aux module, and the
"modules_install_no_aux" build target. In these drivers, simply add
a patch to avoid installing the aux module in "modules_install_no_aux".
Such patches are labeled as:
"<driver-version>-don-t-install-auxiliary-module-on-module.patch"

We can drop the "<driver-version>-no-aux-bus.patch" for ice, as the sole
purpose of this patch was to avoid installing the auxiliary module and
auxiliary_bus.h header file. However, as we are now using
"modules_install_no_aux" instead of "modules_install", no auxiliary
files will be installed for the ice driver anyways.

It is necessary to separate "mandocs_install" and
"modules_install_no_aux" into two separate make commands for the ice
driver. When these targets are executed with the same make command,
which is multithreaded due to %{?_smp_mflags}, there is a conflict
and only the mandocs end up getting installed. This seems to be
because "mandocs_install" in the ice driver calls the "all" build
target and this starts a thread to build the modules. This build
thread and "modules_install_no_aux" thread have some conflict when
scheduled at the same time and "modules_install_no_aux" just doesn't
run.

Change-Id: I36fcfec25cdbe1f524b352ed592155077a48f190
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/20036
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>