tuned: set netdev_queue_count properly
To expand the variables in the conf files, in plugin_net, I created
a new Variable object. However, there is actually a hidden Variable
object in the base plugin class - in base.py. plugin_net.py doesn't
use this object at all, so it flew under the radar. However, we should
use this object to expand the variables instead of the new object,
as the old object has been filled with the necessary properties
already.

Amend the patch to use the base class Variable object, instead of
overwriting with a new Variable object.

Essentially a clone of the commit from dev branch

Change-Id: Ia41172b8a50e409c675917b7b24fbf097308944b
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/20356
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>