glibc: fix locale generation issue
In the existing system, locale-gen.sh requires few extra files to
generate locales properly or else when we run this script it will result
in a login loop due to login daemon crash because of incomplete locale
data.

This patch fixes it by packaging the minimal set of files needed to
generate the basic locale info.

Change-Id: Ie35aa5b1aba86b812867e60336ed92ae42c9a764
Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/20480
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>