kernels: Fix CVE-2022-39189
commit 6cd88243c7e03845a450795e134b488fc2afb736 upstream

If a vCPU is outside guest mode and is scheduled out, it might be in
the process of making a memory access. A problem occurs if another
vCPU uses the PV TLB flush feature during the period when the vCPU is
scheduled out, and a virtual address has already been translated but
has not yet been accessed, because this is equivalent to using a stale
TLB entry.

To avoid this, only report a vCPU as preempted if sure that the guest
is at an instruction boundary. A rescheduling request will be
delivered to the host physical CPU as an external interrupt, so for
simplicity consider any vmexit *not* instruction boundary except for
external interrupts.

Change-Id: I6babef0cf22fcfebb97fef78589c474ecd6890c5
---
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/20494
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>