linux-rt: Fix issues in Guest timer Advancement feature
Code to disable Guest timer Advancement broke in some systems
due to compiler optimizations. Fixed that with memory barriers.

Guest timer advancement will not work in CPUs that do not
support TSC_DEADLINE_TIMER feature with current design. So
we disable the feature in this case.

We are also reducing the severity and frequency of a log in
hrtimer.c
We were printing a log every time we spin for > 1us. 1us spent
spinning does not cause any serious issues, but this message is
flooding dmesg output.
We are increasing the threshold for this message to 5us.
Spinning for 5us is unlikely, but can cause latency issues in RT
systems if a CPU is unavailable for that much time.

Change-Id: Idbb3760a8ebd98e7a1f7e61eba45b7c0a0d9ccbf
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/20543
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>