photon-upgrade: package removal support and error code changes
To support some peculiar upgrade scenarios, like that of appliances.
Now photon-ugrade package provides control to the user to specify
the list of packages they want to be removed pre/post upgrade.

Additionally, exit codes 11 and 22 were employed to signify to the
caller of the package to indicate that the photon-upgrade.sh can be
retried.

Change-Id: I03686d4f54bdf98f42e385924f48862e51957967
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/20575
Tested-by: Michelle Wang <michellew@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>