tuned: skip smp affinity verification of irq 0 on x86
Tuned does not set irqaffinity on the kernel command line, causing
verification of the realtime profile to fail when IRQ 0 is present,
as the affinity of IRQ 0 cannot be changed after boot.

On x86 machines, irq 0 is reserved for the system timer, and won't be
used for anything else anywhere in the kernel. The only exception is if
some other device driver hijacks the physical interrupt line for it's own
purposes, but that seems very unlikely and maybe impossible unless the
timer hardware is missing for some reason.

irq 0 can only be balanced with irqaffinity on the command line, which sets
the default cpu mask for all irqs when they are initially allocated; it
remains impossible to modify this for irq 0 after boot. Also, these interrupts
should always be delivered to the boot CPU (cpu 0) in the vast majority of systems.
Either way, as irq 0 cannot be threaded/balanced by the kernel,
these interrupts should only be handled by the CPU where the interrupt is delivered.

Photon RT is configured with CONFIG_NO_HZ as well, so the main system timer
interrupt (IRQ 0) is disabled after boot time anyways and won't be used on a
running system, so setting the affinity is also pointless here.

In order to properly verify realtime settings, add a patch to skip
verification of the SMP affinity of irq 0, as it seems safe to do so on
x86.

Change-Id: I57fc63c2b2f84db97fe4ece7f821ff15400e627f
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/20644
Reviewed-by: Srivatsa S. Bhat <srivatsab@vmware.com>
Tested-by: Srivatsa S. Bhat <srivatsab@vmware.com>