linux-esx: add Intel i915 backport driver
It comes from 3rd party github project
https://github.com/intel-gpu/intel-gpu-i915-backports/

In addition we apply huge patch to allow building of headless
(no display support) version of i915.

Applying also series of patches to reduse dependencies tree support
from linux-esx kernel to keep it small and fast.
i915 still requires additional kernel support such as: MTD, MFD, AGP.
It was enabled in config-esx.

As an addition to this change, we also mark MTRR as always enabled.
It allows to use PAT in CRX instances.

Change-Id: I7f340100faafa66eb8e3e8c8bbf744dd6a12cdaf
Signed-off-by: Alexey Makhalov <amakhalov@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/21255
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-by: Keerthana K <keerthanak@vmware.com>