photon-upgrade: fix upgrade failure using public repo
During the upgrade process of 3.0 to 5.0, the installed ansible pkg
was getting downgraded on removal of python3-pycrypto briging in
many python2 packages which were already reemoved as part of removal
of deprecated packaes. The fix is a workaround to add ansible to the
replaced package map so that it gets uninstalled before upgrading
to 5.0 and gets reinstalled after upgrading OS.

This issue was seen only in public repositories. Appliances would not
run into this issue.

Change-Id: If8c068102eca9278084153590ba33c527538e7f2
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/21018
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-by: Tapas Kundu <tkundu@vmware.com>