falco: build Go components locally
Falco plugins (.so files in the falco rpm) are not built during our build
process, but are instead just consumed as binaries that are built long ago,
potentially with an outdated Go compiler, which may allow some CVEs to linger
even after Go is updated in Photon.

falcoctl is also consumed the same as the plugins, except as a standalone binary.

Patch the build process to build these plugins/falcoctl from source with the updated,
hopefully safer Photon toolchain instead of consuming old prebuilt binaries.

Change-Id: I7f3d2f07b625f6ad6a138b451344cbf22451e0ed
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/21219
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>