falco: build plugins locally
Falco plugins (.so files in the falco rpm) are not built during our build
process, but are instead just consumed as binaries that are built long ago,
potentially with an outdated Go compiler, which may allow some CVEs to linger
even after Go is updated in Photon.

Patch the build process to build these plugins from source with the updated,
hopefully safer Photon toolchain instead of consuming old prebuilt binaries.

Change-Id: Ia0488ca5e8eaac865dd27e615ed791d6dc2a23c1
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/21222
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Shreenidhi Shedi <sshedi@vmware.com>