dracut: fix an issue with finding installed kernels during initrd creation
In the current system, any valid directory under /lib/modules is
getting considered as an installed kernel entry and if a dummy directory
is present under /lib/modules, initrd generation will fail and system
upgrade fails.

This fix makes correction to this logic and now we consider only the
kernels which have a corresponding vmlinuz under /boot.

Change-Id: I5e44873c043f9bc8295a92732d6a5fdfd2f79385
Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/21278
Reviewed-by: Keerthana K <keerthanak@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>