dracut: fix an issue with finding installed kernels during initrd creation
In the current system, any valid directory under /lib/modules is
getting considered as an installed kernel entry and if a dummy directory
is present under /lib/modules, initrd generation will fail and system
upgrade fails.

This fix makes correction to this logic and now we consider only the
kernels which have a corresponding vmlinuz under /boot.

Change-Id: I27125c1041eabcea0bb18c942b9dfb0e9a3677ee
Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/21299
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>