stalld: Move logging to dedicated thread
All logging in stalld is synchronous. Moreover stalld operates in an
environment where system hangs due to various reasons are common.
We have encountered several instances where IO is stuck across the system.
As a consequence stalld also gets stuck in uninterruptible sleep state.

We are moving all logging to a dedicated thread, so that even if IO is
blocked, stalld can continue boosting starving threads.

Change-Id: I5b2dd19754a11c5f3fe9289f8d6fbba25890aa50
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/21319
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>