linux-esx: Revert CRYPTO_TEST (tcrypt) to 'm'
tcrypt test fails in deflate only on PowerEdge R650 servers
during bootup in fips mode.

CONFIG_CRYPTO_TEST (tcrypt) was set to 'y' accidentally in linux-esx
with this commit 47d1b0ce19fedfb0d8b682436a805f72020e935e and
CONFIG_DEFLATE is 'm'.

tcrypt is not needed as 'y' as this is not part of canister.
But it passes on other servers without any issue and gerrit test
passed as well. Module loading of deflate fails during bootup
only on R650 servers and subsequently fails the tcrypt test.

Change-Id: Iae4df91648c23a14fb64d050253d8ca4daa1daf9
Signed-off-by: Keerthana K <keerthanak@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/22018
Reviewed-by: Vamsi Krishna Brahmajosyula <vbrahmajosyula@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/22513