sqlite: patched CVE-2023-7104
The upstream patch: https://sqlite.org/src/info/0e4e7a05c4204b47
is not directly applicable to the source that we are consuming because
Expected source: https://www.sqlite.org/src/tarball/sqlite.tar.gz?r=release
But we are consuming: https://sqlite.org/2022/sqlite-autoconf-%{sourcever}.tar.gz

So the patch had to be modified.
The upstream patch involves 2 code changes.

Change1: code change to mitigate the vulnerability - Implemented in our patch

Change2: test code - Not implemented in our patch as the required test code to be patched is not found in the source we are consuming

Change-Id: Ie02b8cb1a8ceaf305d3ab69a8d61994e2f333823
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/23040
Tested-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>