stalld: Add two features to stalld
Threaded logging:
All logging in stalld is synchronous. Moreover stalld operates in an
environment where system hangs due to various reasons are common.
We have encountered several instances where IO is stuck across the system.
As a consequence stalld also gets stuck in uninterruptible sleep state.

We are moving all logging to a dedicated thread, so that even if IO is
blocked, stalld can continue boosting starving threads

Boost and Migrate:
Stalld is intended to keep a system stable. In order to do that it monitors
threads that are starving, and then boosts their priority using the deadline
scheduler. Currently any starving thread is boosted, including user space
threads. This pactice however leads to unintended consequences in misconfigured
systems.

In practice, we have seen instances where multiple user threads are starving on
a CPU (which is also running one or more high priority cpu hungry task), along
with kernel threads. The lower priority user threads, as well as kernel threads
are entirely dependent on stalld to make progress.

In such cases, stalld spends significant time boosting user threads. As a result
kernel thread spend more time being starved and unable to make progress. This
leads to systemwide issues, including system hangs, thus defeating the purpose of
stalld

With this patch, we change stalld's behaviour to boost kernel threads and select
non CPU agnostic usertasks only. A user task that is CPU agnostic, ie. one that
can be moved to a different CPU will be migrated to a CPU that is not as busy.
We have added a config to provide a whitelist of non CPU agnostic user process
and threads.

Change-Id: Ic3558afbbfe5be921e85ad6a1e1318116d8c2bc6
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/23213
Reviewed-by: Ankit Jain <ankitja@vmware.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Dweep Advani <dadvani@vmware.com>