photon-upgrade: avoid coredumps during upgrade
Sometimes coredump may happen during upgrade when an underlying shared
library gets updated which is being used by a running process.

The existence of coredumps fails some of the appliances CI/CD
pipelines. Hence, during the upgrade/update, disable generation of
coredumps.

This change also reverts the change of logging to a logfile as
the caller is expected to log the output as per their choice.

Change-Id: I4351a5437a72ef1e341687514373fbc8fdbc88e6
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/23981
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Tapas Kundu <tapas.kundu@broadcom.com>
Reviewed-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>