glibc: fix locale generation issue
In the existing system, locale-gen.sh requires few extra files to
generate locales properly or else when we run this script it will result
in a login loop due to login daemon crash because of incomplete locale
data.

This patch fixes it by packaging the minimal set of files needed to
generate the basic locale info.

This is the backport of fix:
https://github.com/vmware/photon/commit/b1b92448d8ef2fd6c42217ac6f7adbf74030e3d3
From Ph5.

Change-Id: I988fd2d3d62f146368ef6b1ccd341024771914ca
Signed-off-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/24158
Reviewed-by: Srinidhi Rao <srinidhi.rao@broadcom.com>
Reviewed-by: Vamsi Krishna Brahmajosyula <vamsi-krishna.brahmajosyula@broadcom.com>