package-builder/SpecParser.py: stop at if condition while fetching description
Continuing to read after seeing an if condition might result in wrong
data. Example:

```
%description
line1
line2
...

%if 0%{?condition}
%package pkg
...
%endif
```

Here if we don't stop at if condition, parser will consider `pkg` as a
package to build even if `condition` is false. One such instace is
`linux-fips-canister` package in linux.spec

The downside of this is, we might miss some description strings if we
use a if condition in the description itself. But this is a rare
occurence and missing some data from description field doesn't harm.

Change-Id: Ibb5cec2fce63862b165fdfe5ac0fa59e2bd583dd
Signed-off-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/24358
Reviewed-by: <mukul.sikka@broadcom.com>
Reviewed-by: Alexey Makhalov <amakhalov@vmware.com>
Reviewed-by: Keerthana K <keerthana.kalyanasundaram@broadcom.com>