kernels: fix CVE-2024-42080, CVE-2021-47188
CONFIG_INFINIBAND is not set, so CVE-2024-42080 is not
applicable to linux-esx.

CONFIG_SCSI_UFSHCD is not set in aws/esx, so CVE-2021-47188
is not applicable to those flavors.

The more correct fix for CVE-2021-47188 would be to
backport these two patches:
https://lore.kernel.org/all/20211203231950.193369-13-bvanassche@acm.org/
https://lore.kernel.org/all/20211130233324.1402448-14-bvanassche@acm.org/

However, backporting these is deceptively complicated given the change in
context and many other missing backports. So, let's just use the original,
simple fix.

Change-Id: I3f5406adf22f7a51bfc708a8ef274efa1fbbbd90
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/24455
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-by: Ajay Kaher <ajay.kaher@broadcom.com>
Reviewed-by: Alexey Makhalov <alexey.makhalov@broadcom.com>