Sandbox: refactor putFiles() and observer
Sandbox.putFiles(files, dest, ...) now uses a similar semantic as
subprocess.run(args, ...), where files can either be a list of
files, or a single string. For a list of files, dest will be the
directory the files get copied to. For a single file, dest is the
actual path of destination file.

Observer is now a separate class type. The logic is detached from
nspawn sandbox type and can be applied to other types now. The
class is intended to be used with a context manager to ease the
lifecycle management of observer container.

Misc. other minor fixes:
- tempfile.NamedTemporaryFile needs to be flushed for the file
  content to be fully visible for external programs
- Define network_required for package that needs it
- When network_required is defined, ca-certificates should be
  installed for the observer append its own ca-cert.
- New print-java-home.java script to determine JAVA_HOME

Change-Id: I35b4413116963aff6e74164e2323f6094adb453e
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/c/photon/+/24537
Reviewed-by: Alexey Makhalov <alexey.makhalov@broadcom.com>
Reviewed-by: Bo Gan <bo.gan@broadcom.com>
Tested-by: gerrit-photon <photon-checkins@vmware.com>
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/24768
Reviewed-by: Ajay Kaher <ajay.kaher@broadcom.com>
Reviewed-by: Vamsi Krishna Brahmajosyula <vamsi-krishna.brahmajosyula@broadcom.com>
Tested-by: Ajay Kaher <ajay.kaher@broadcom.com>