kernels: add jitterentropy rng proxy
Add a built-in kernel module to act as a proxy,
managing some arbitrary number of underlying
jitterentropy instances.

The proxy takes over the "jitterentropy_rng" handle,
so any requests to jitterentropy should come through it.

This is meant to address some performance issues with
jitterentropy, specifically the fact that a new
memory chunk is dynamically allocated when a new
jitterentropy instance is instantiated. For example,
this happens anytime a new crypto socket is bound to
"jitterentropy_rng". Theoretically, there is no limit
on the number of jitterentropy instances that can exist
at any time, and so it's possible to run into OOM or
fragmentation issues.

The proxy fixes these issues and improves memory usage by:
1. Dynamic allocation upon request only
2. Deallocation when no longer immediately needed
3. Limiting the max number of jitterentropy instances

Instead of allocating upon handle/socket opening/closing,
allocation/deallocation is done at request time only. After
the first request, we'll keep one instance allocated at all
times, however the rest of the instances will be deallocated
if there are no pending requests.

The max number of jitterentropy instances can be controlled by
"jitterentropy_rng_proxy.max_jent_insts". The default max is 1.

Change-Id: I993b9e9493d662c0c6e1b7ab1d7a3f5e42db2529
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/24687
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>
Reviewed-by: Srinidhi Rao <srinidhi.rao@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>