tools/scripts: Add helper script to validate SPDX license expressions
As we are moving to replace our more casual license strings with
SPDX license identifiers, we should also be careful that
the new strings are parseable as valid SPDX expressions.

This commit adds a helper script for developers which can
validate SPDX expressions either from command line input or
from a file, such as our license.txt or .spec files.

Change-Id: Ia349a37f3e1dc599e8b36c74f1e11c9cf9777d3b
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/25288
Tested-by: <mukul.sikka@broadcom.com>
Reviewed-by: <mukul.sikka@broadcom.com>
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>