tools/scripts: Add helper script to validate SPDX license expressions
As we are moving to replace our more casual license strings with
SPDX license identifiers, we should also be careful that
the new strings are parseable as valid SPDX expressions.

This commit adds a helper script for developers which can
validate SPDX expressions either from command line input or
from a file, such as our license.txt or .spec files.

Change-Id: I7535d8d4f6fe9517d7c2a871e0c6c1144749d066
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/24790
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>