kernels: Fix structure aead_geniv_ctx compatibility issue
Issue:

structure aead_geniv_ctx has member spinlock_t lock. This member
is built as void *lock in fips_canister.o whereas in rest of the kernel
it is compiled as spinlock_t lock. We have made it as void *lock
to avoid structure differences between fips_canister.o and kernel (especially
RT). This causes an issue when used with echainiv generator which is not part
of fips_canister.o.

As this lock variable is just initialized and not used anywhere in the IV
generator code or geniv.c, we can safely assume that this variable is
not used anywhere.

Also, there is a memory leak on the lock variable. This member lock is
allocated in fips_canister_wrapper but there is no code added for freeing
the memory.

Fix:

Change this member as void * and rename it as reserved for the
rest of the kernel.

Since this member lock is not used anywhere, we can remove the alloc and
initialization code of this spinlock and return a dummy variable to avoid
the memory leak.

Updates to check_fips_canister_compatibility GDB script:

1. Compare symbols with offset and size using ptype/o
2. Maintain list of structures and its definition that are different between
   canister and vmlinux by offset and size.
3. Though these structures are different between vmlinux and canister because of
   size of variables, the overall size has not changed, so we can ignore these
   structure differences between canister and vmlinux. Added a new check to fail
   the build if there is change in the definition of these known symbols in
   vmlinux in future.

Change-Id: Iaae8c3a762fb0560bf6bed46a5d98ef327813b7c
Signed-off-by: Keerthana K <keerthana.kalyanasundaram@broadcom.com>
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/24802
Reviewed-by: Alexey Makhalov <alexey.makhalov@broadcom.com>
Reviewed-by: Srinidhi Rao <srinidhi.rao@broadcom.com>
Reviewed-by: Ajay Kaher <ajay.kaher@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>