kernels: fix UDF filesystem directory traversal regression
Starting from linux-6.1.114, UDF or ISO disks in general terms showed
regression when the file was mounted and when ls -lR or find was run on
the mounted location.

```
mount -o loop file.iso /mnt
find /mnt
```
When the above find command was run, the following dmesg logs are seen.

```
[   51.682153] UDF-fs: error (device loop0): udf_verify_fi: directory (ino 281) has too big (2088) entry at pos 2032
[   51.682283] UDF-fs: error (device loop0): udf_fiiter_advance_blk: extent after position 2032 not allocated in directory (ino 281)
[   51.682525] UDF-fs: error (device loop0): udf_verify_fi: directory (ino 281) has too big (2088) entry at pos 2032
[   51.682655] UDF-fs: error (device loop0): udf_fiiter_advance_blk: extent after position 2032 not allocated in directory (ino 281)
```

This does not happen with all ISO files but only few ISOs result in this.
And some appliance teams got affected by this regression.

Change-Id: If2c21adfb7124361b55e1bddb5a66b69bb3cb90d
Signed-off-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/24908
Reviewed-by: Alexey Makhalov <alexey.makhalov@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: Bo Gan <bo.gan@broadcom.com>