cups: update to 2.4.11
There are several CVEs targeting the CUPS software,
some of them for various subpackages such as cups-browsed,
or libppd. These subpackages often borrow lots of code
from the mainline CUPS package, causing CVEs to be
theoretically applicable in both places.

These CVEs can be combined and exploited for remote command execution
as described in
https://www.evilsocket.net/2024/09/26/Attacking-UNIX-systems-via-CUPS-Part-I/

These CVEs for CUPS and various CUPS related packages include:
 - CVE-2024-47176
 - CVE-2024-47076
 - CVE-2024-47175
 - CVE-2024-47177

While Photon is *NOT* at risk of this particular exploit chain,
because we don't have the cups-browsed service, CVEs
such as CVE-2024-47175 which applies to libppd also
affects the same code in mainline CUPS and should be patched.

There are 5 commits needed to remediate this exploit in mainline CUPS,
as described in
https://www.openwall.com/lists/oss-security/2024/09/27/3

Update to the latest subversion 2.4.11 in order to consume
these fixes.

Change-Id: Ieff8b832dfeb1004c1dcd3b7dd93b0c834a88ffd
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/24930
Reviewed-by: Harinadh Dommaraju <harinadh.dommaraju@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: Michelle Wang <michelle.wang@broadcom.com>
Reviewed-by: Prashant Singh Chauhan <prashant.singh-chauhan@broadcom.com>