gcc: add callback for plugin before struct layout
This allows us to inject modifications for structures
right before the final layout of the structure.

This is needed because the alternative, PLUGIN_FINISH_TYPE,
happens too late and the DWARF debuginfo is already written.

We can use this plugin callback to keep the debuginfo in line
with the final structure layout.

Change-Id: I042ed47c5ad2736629ae26c16ed536feb60f04fc
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/25770
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>
Tested-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-by: Ankit Jain <ankit-aj.jain@broadcom.com>