linux: introduce FIPS canister GCC plugins
This commit introduces two new GCC plugins which are meant
to help future-proof and safeguard the interface between the Linux
kernel and the static frozen-in-time FIPS canister binary.

pad_canister_struct.so - This plugin adds 8 bytes of padding
to each whitelisted structure. This is ONLY run at canister
build time. Extra padding for canister structures helps
to future proof binary compatibility with the upstream kernel.

match_canister_strucs.so - This plugin reads the structure data
from the FIPS canister, and adapts the kernel structures to match.
The aforementioned padding added to the FIPS canister structures
can be used by this plugin to absorb new structure members.
For speed/efficiency, structure data is cached after the first
plugin invocation and mmap'd in to subsequent invocations.

match_canister_structs.so does two main operations:
  1. Match ordering of structure members between kernel and canister
     Any new fields, if possible, are placed in the open padding space
     at the END of the structure, to preserve the offsets of canister
     known fields. In the case of flexible array members, the padding
     is placed before the flexible array member, and the new field
     will replace the padding there.
  2. Match size of structure between kernel and canister - inject
     padding to ensure kernel and canister are the same size. This plugin
     does not bother with implicit padding, and so the size of padding fields
     may not match between canister and kernel, even if the size ends
     up being the same (since the canister padding is always 8 bytes).
  3. Safety checks - offset validation, member matching, size checks, etc.

One side effect of these plugins is that because we are messing
with structure layouts, we need to restrict constructor initialization
to designated initializers only.

Future development:
1. Match struct member type names with the versions in DWARF debug info,
   in order to implement a fully functional structure member type check.
   This is the last major hurdle to replacing the struct comparator check
   which is run post-build.

Notes:
- Structure comparator tests will have to be updated when the new canister
  is consumed, as the padding fields will introduce differences.
- Self tests are included, and should showcase the intended behavior in a
  variety of situations. Please run when doing any new developement.
- linux-devel requires libdwarves, as this is linked to the plugin .so's

Change-Id: If59cf14a40ef761345790d54cc3baf59638128f1
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/25766
Reviewed-by: Vamsi Krishna Brahmajosyula <vamsi-krishna.brahmajosyula@broadcom.com>
Reviewed-by: Ankit Jain <ankit-aj.jain@broadcom.com>
Tested-by: Vamsi Krishna Brahmajosyula <vamsi-krishna.brahmajosyula@broadcom.com>